Refactor: Convert User Profile tab to Nuxt UI
Replace legacy i18n attributes with $t(), gui_box with UiBox component,
globalThis.confirm with useDialog composable, and custom CSS modal with
native dialog element. Rename to UserProfileTab.vue for consistency.
